Modify main.cpp

+++ main.cpp
@@ -2,6 +2,7 @@
 using namespace std;
 
 pthread_t thread[PTHREAD_NUM];
+pthread_t session_thread[PTHREAD_NUM];
 pthread_cond_t cond = PTHREAD_COND_INITIALIZER;
 pthread_mutex_t mutex = PTHREAD_MUTEX_INITIALIZER;
 int num_of_parameter = 0;
@@ -9,35 +10,34 @@ pid_t pid[SESSION_NUM];
 pcap_t* handle;
 
 void terminate_Process(int sig){
-    for(int i = 0; i < num_of_parameter; i++){
-        kill(pid[i], SIGKILL);
-        pthread_join(thread[i], NULL);
+    if(sig == SIGINT){
+        for(int i = 0; i < num_of_parameter; i++){
+            kill(pid[i], SIGKILL);
+            pthread_join(thread[i], NULL);
+            pthread_join(session_thread[i], NULL);
+        }
+        pcap_close(handle);
     }
-    pcap_close(handle);
 }
 
 char* make_Parameter_REQ(int num, char* at_ip_addr, char* at_mac_addr, char* ip, pcap_t* handles){
     struct Parameter_Pthread* pt2 = (struct Parameter_Pthread*)malloc(sizeof(struct Parameter_Pthread));
-    char* temp_mac = (char*)malloc(sizeof(char) * 12);
+    char* temp_mac = (char*)malloc(sizeof(char) * 20);
     if(temp_mac == NULL){ perror("mac malloc error"); exit(1); }
 
-    pt2->sip = (char*)malloc(strlen(at_ip_addr));
-    if(pt2->sip == NULL){ perror("pt2->sip malloc error"); exit(1); }
-    pt2->smac = (char*)malloc(strlen(at_mac_addr));
-    if(pt2->smac == NULL){ perror("pt2->smac malloc error"); exit(1); }
-
-    pt2->tip = (char*)malloc(strlen(ip));
-    if(pt2->tip == NULL){ perror("pt2->tip malloc error"); exit(1); }
-    pt2->tmac = (char*)malloc(strlen(temp_mac));
-    if(pt2->tmac == NULL){ perror("pt2->tmac malloc error"); exit(1); }
-
-    strncpy(pt2->sip, at_ip_addr, strlen(at_ip_addr));
-    strncpy(pt2->smac, at_mac_addr, strlen(at_mac_addr));
-    strncpy(pt2->tip, ip, strlen(ip));
+    strncpy(pt2->sip, at_ip_addr, 16);
+    strncpy(pt2->smac, at_mac_addr, 12);
+    pt2->smac[13] = '\0';
+    strncpy(pt2->tip, ip, 16);
     //strncpy(pt2->tmac, temp_mac, strlen(temp_mac));
     pt2->handle = handles;
     pt2->session_Number = num;
-    pt2->smac[12] = '\0';
+    //pt2->sip[strlen(pt2->sip)] = '\0';
+    //pt2->tip[strlen(pt2->tip)] = '\0';
+
+    printf("pt2->sip : %s\n", pt2->sip);
+    printf("pt2->smac : %s\n", pt2->smac);
+    printf("pt2->tip : %s\n", pt2->tip);
 
     pthread_create(&thread[num], NULL, find_Mac, (void*)(pt2));
     pthread_join(thread[num], (void**)(&temp_mac));
@@ -58,13 +58,14 @@ int main(int argc, char* argv[]) {
     }
 
     num_of_parameter = (argc-1)/2;
-    if(num_of_parameter > 20){
+    if(num_of_parameter > SESSION_NUM){
         perror("Overflow of Session!");
         exit(0);
     }
     printf("[ NUM_OF_PRARMETER : %d ]\n", num_of_parameter);
 
     char* dev = argv[1];
+    printf("[ dev : %s ]\n", dev);
 
 // ########## Set sigaction fuction for SIGINT(ctrl + c) ##########
     struct sigaction sa;
@@ -90,15 +91,15 @@ int main(int argc, char* argv[]) {
 // ########################################
 
 // ########## Find my IP & MAC address ##########
-    struct Info_mymac* IM = (struct Info_mymac*)malloc(sizeof(struct Info_mymac));
-    char* a_mac_addr = (char*)malloc(12);
+    char* a_mac_addr = (char*)malloc(13);
     if(a_mac_addr == NULL){ perror("a_mac_addr malloc error"); exit(1); }
-    char* s_ip_addr = (char*)malloc(sizeof(uint32_t));
+    char* s_ip_addr = (char*)malloc(16);
     if(s_ip_addr == NULL){ perror("s_ip_addr malloc error"); exit(1); }
-    IM = find_My_Mac();
-    strncpy(a_mac_addr, IM->my_mac, strlen(IM->my_mac));
-    strncpy(s_ip_addr, IM->my_ip, strlen(IM->my_ip));
-    a_mac_addr[12] = '\0';
+    struct Info_mymac* IM = find_My_Mac();
+    strncpy(a_mac_addr, IM->my_mac, 12);
+    a_mac_addr[13] = '\0';
+    strncpy(s_ip_addr, IM->my_ip, 16);
+    a_mac_addr[strlen(a_mac_addr)] = '\0';
     s_ip_addr[strlen(s_ip_addr)] = '\0';
     printf("[ Success to find my mac address : %s ]\n", a_mac_addr);
 // ########################################
@@ -112,21 +113,22 @@ int main(int argc, char* argv[]) {
     Session s[SESSION_NUM];
     for(int i = 0; i < num_of_parameter; i++){
         // Find mac using pcap_sendpacket
-        smac[i] = (char*)malloc(sizeof(char) * 40);
+        smac[i] = (char*)malloc(13);
         if(smac[i] == NULL){ perror("smac malloc error"); exit(1); }
-        char* t_smac = (char*)malloc(sizeof(char) * 40);
+        char* t_smac = (char*)malloc(13);
         if(t_smac == NULL){ perror("t_mac malloc error"); exit(1); }
         t_smac = make_Parameter_REQ(i+1, s_ip_addr, a_mac_addr, argv[cnt], handle);
         smac[i] = t_smac;
+        smac[i][13] = '\0';
         printf("[ < %d > Success to find < %s > mac address : %s ]\n", i+1, argv[cnt], smac[i]);
 
-        dmac[i] = (char*)malloc(sizeof(char) * 40);
+        dmac[i] = (char*)malloc(13);
         if(dmac[i] == NULL){ perror("dmac malloc error"); exit(1); }
-        char* t_dmac = (char*)malloc(sizeof(char) * 40);
+        char* t_dmac = (char*)malloc(13);
         if(t_dmac == NULL){ perror("t_mac malloc error"); exit(1); }
-        printf("argv[%d] : %s\n", cnt+1, argv[cnt+1]);
         t_dmac = make_Parameter_REQ(i+1, s_ip_addr, a_mac_addr, argv[cnt+1], handle);
         dmac[i] = t_dmac;
+        dmac[i][13] = '\0';
         printf("[ < %d > Success to find < %s > mac address : %s ]\n", i+1, argv[cnt+1], dmac[i]);
         printf("%d smac : %s\n", i, smac[i]);
         printf("%d dmac : %s\n", i, dmac[i]);
@@ -138,34 +140,28 @@ int main(int argc, char* argv[]) {
     for(int i = 0; i < num_of_parameter; i++){
         pt[i] = (struct Parameter_Pthread*)malloc(sizeof(struct Parameter_Pthread));
         // ########## Make contents of Attack parameters ##########
-        pt[i]->sip = (char*)malloc(strlen(argv[cnt]));
-        if(pt[i]->sip == NULL){ perror("pt->sip malloc error"); exit(1); }
-        pt[i]->smac = (char*)malloc(strlen(a_mac_addr));
-        if(pt[i]->smac == NULL){ perror("pt->smac malloc error"); exit(1); }
-        pt[i]->tip = (char*)malloc(strlen(argv[cnt+1]));
-        if(pt[i]->tip == NULL){ perror("pt->tip malloc error"); exit(1); }
-        pt[i]->tmac = (char*)malloc(strlen(smac[i]));
-        if(pt[i]->tmac == NULL){ perror("pt->tmac malloc error"); exit(1); }
-        strncpy(pt[i]->sip, argv[cnt+1], strlen(argv[cnt+1]));
-        strncpy(pt[i]->smac, a_mac_addr, strlen(a_mac_addr));
-        strncpy(pt[i]->tip, argv[cnt], strlen(argv[cnt]));
-        strncpy(pt[i]->tmac, smac[i], strlen(smac[i]));
+        strncpy(pt[i]->sip, argv[cnt+1], 16);
+        strncpy(pt[i]->smac, a_mac_addr, 12);
+        pt[i]->smac[13] = '\0';
+
+        strncpy(pt[i]->tip, argv[cnt], 16);
+        strncpy(pt[i]->tmac, smac[i], 12);
+        pt[i]->tmac[13] = '\0';
+
         pt[i]->handle = handle;
         pt[i]->session_Number = i+1;
         // ########################################
 
+        printf("main pass\n");
         // ########## Attack ##########
         s[i].set(i+1, smac[i], argv[cnt], dmac[i], argv[cnt+1], handle, a_mac_addr, pt[i]);
         // ########################################
         cnt += 2;
     }
 
-    while(1){
-        for(int i = 0; i < num_of_parameter; i++)
-            s[i].handle_session();
+    for(int i = 0; i < num_of_parameter; i++){
+       s[i].handle_session();
     }
-// @@@@@@@@@@ @@@@@@@@@@ @@@@@@@@@@ @@@@@@@@@@
-// @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
 }
